Cloned project - this is SteamVR version
I've split my dissertation into 2 versions, one that will purely run with steamvr on unity, and the other that will focus on integrating VRChat into the project. For now, this repo will be where I perform all the changes and testing. Once this is complete, I will move over all the changes to the VRChat repo, and begin the process of converting my project into a format that works with VRChat.